move activities to org.cru.godtools package
