+++ app/src/main/java/org/cru/godtools/activity/BaseActivity.java
@@ -1,4 +1,4 @@
-package org.keynote.godtools.android.activity;
+package org.cru.godtools.activity;
 
 import android.content.ActivityNotFoundException;
 import android.content.Intent;
@@ -20,7 +20,6 @@ import com.google.common.base.Objects;
 import com.google.common.base.Strings;
 
 import org.ccci.gto.android.common.compat.util.LocaleCompat;
-import org.cru.godtools.activity.AboutActivity;
 import org.cru.godtools.base.Settings;
 import org.keynote.godtools.android.BuildConfig;
 import org.keynote.godtools.android.R;
