+++ app/src/main/java/org/cru/godtools/activity/LanguageSettingsActivity.java
@@ -1,4 +1,4 @@
-package org.keynote.godtools.android.activity;
+package org.cru.godtools.activity;
 
 import android.content.Context;
 import android.content.Intent;
@@ -8,21 +8,17 @@ import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
 import android.support.v4.app.Fragment;
 import android.support.v4.app.FragmentManager;
-import android.view.Menu;
 
 import org.keynote.godtools.android.R;
-import org.keynote.godtools.android.fragment.ToolsFragment;
-import org.keynote.godtools.android.model.Tool;
+import org.keynote.godtools.android.fragment.LanguageSettingsFragment;
 
-import java.util.Locale;
+import static org.cru.godtools.analytics.AnalyticsService.SCREEN_LANGUAGE_SETTINGS;
 
-import static org.cru.godtools.analytics.AnalyticsService.SCREEN_ADD_TOOLS;
-
-public class AddToolsActivity extends BaseActivity implements ToolsFragment.Callbacks {
+public class LanguageSettingsActivity extends BaseActivity {
     private static final String TAG_MAIN_FRAGMENT = "mainFragment";
 
     public static void start(@NonNull final Context context) {
-        context.startActivity(new Intent(context, AddToolsActivity.class));
+        context.startActivity(new Intent(context, LanguageSettingsActivity.class));
     }
 
     /* BEGIN lifecycle */
@@ -34,12 +30,6 @@ public class AddToolsActivity extends BaseActivity implements ToolsFragment.Call
     }
 
     @Override
-    public boolean onCreateOptionsMenu(@NonNull final Menu menu) {
-        getMenuInflater().inflate(R.menu.activity_add_tools, menu);
-        return true;
-    }
-
-    @Override
     protected void onStart() {
         super.onStart();
         loadInitialFragmentIfNeeded();
@@ -48,22 +38,7 @@ public class AddToolsActivity extends BaseActivity implements ToolsFragment.Call
     @Override
     protected void onResume() {
         super.onResume();
-        mAnalytics.onTrackScreen(SCREEN_ADD_TOOLS);
-    }
-
-    @Override
-    public void onToolSelect(@Nullable final String code, @NonNull final Tool.Type type, final Locale... languages) {
-        ToolDetailsActivity.start(this, code);
-    }
-
-    @Override
-    public void onToolInfo(@Nullable final String code) {
-        ToolDetailsActivity.start(this, code);
-    }
-
-    @Override
-    public void onNoToolsAvailableAction() {
-        finish();
+        mAnalytics.onTrackScreen(SCREEN_LANGUAGE_SETTINGS);
     }
 
     /* END lifecycle */
@@ -80,7 +55,12 @@ public class AddToolsActivity extends BaseActivity implements ToolsFragment.Call
 
         // update the displayed fragment
         fm.beginTransaction()
-                .replace(R.id.frame, ToolsFragment.newAvailableInstance(), TAG_MAIN_FRAGMENT)
+                .replace(R.id.frame, LanguageSettingsFragment.newInstance(), TAG_MAIN_FRAGMENT)
                 .commit();
+    }
+
+    @Override
+    public void supportNavigateUpTo(@NonNull final Intent upIntent) {
+        finish();
     }
 }
